+++ gamedori/src/gamedori/servlet/member/MemberFindIdServlet.java
@@ -1,4 +1,4 @@
-package servlet.member;
+package gamedori.servlet.member;
 
 import java.io.IOException;
 
@@ -11,31 +11,31 @@ import javax.servlet.http.HttpServletResponse;
 import gamedori.beans.dao.MemberDao;
 import gamedori.beans.dto.MemberDto;
 
-@WebServlet(urlPatterns = "/guest/find_pw.do")
-public class MemberFindPwServlet extends HttpServlet{
+@WebServlet(urlPatterns = "/guest/find_id.do")
+public class MemberFindIdServlet extends HttpServlet{
 	@Override
 	protected void doPost(HttpServletRequest req, HttpServletResponse resp) throws ServletException, IOException {
 		try {
 			//입력 : member_nick , member_phone , member_birth  -----> MemberDto
 			req.setCharacterEncoding("UTF-8");
 			MemberDto mdto = new MemberDto();
-			mdto.setMember_id(req.getParameter("member_id"));
+			mdto.setMember_nick(req.getParameter("member_nick"));
 			mdto.setMember_phone(req.getParameter("member_phone"));
 			mdto.setMember_name(req.getParameter("member_name"));
 			
 			//처리
 			MemberDao mdao = new MemberDao();
-			String member_pw = mdao.findPw(mdto);
+			String member_id = mdao.findId(mdto);
 			
 			//출력
-			if(member_pw != null) {//결과가 있으면(정보가 맞다면)
+			if(member_id != null) {//결과가 있으면(정보가 맞다면)
 //				resp.sendRedirect("find_id_result.jsp?member_id="+member_id);
 				
-				req.getSession().setAttribute("member_pw", member_pw);
-				resp.sendRedirect("find_pw_result.jsp");
+				req.getSession().setAttribute("member_id", member_id);
+				resp.sendRedirect("find_id_result.jsp");
 			}
 			else {//결과가 없으면(정보가 맞지 않으면)
-				resp.sendRedirect("find_pw.jsp?error");
+				resp.sendRedirect("find_id.jsp?error");
 			}
 		}
 		catch(Exception e) {
