+++ gamedori/src/gamedori/servlet/member/MemberJoinServlet.java
@@ -1,4 +1,4 @@
-package servlet.member;
+package gamedori.servlet.member;
 
 import java.io.IOException;
