+++ gamedori/src/gamedori/servlet/qna/QnaEditServlet.java
@@ -1,4 +1,4 @@
-package servlet.qna;
+package gamedori.servlet.qna;
 
 import java.io.IOException;
